Updated OAS type to take into account other versions.
